requirements.txt
# VoiceCraft Dependencies

# AI/ML
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Text Processing
spacy>=3.7.0
nltk>=3.8.0
textstat>=0.7.0
textblob>=0.17.0

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
flask>=3.0.0  # For Slack bot server

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0  # Beautiful CLI output
click>=8.1.0  # CLI framework
requests>=2.31.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# File Processing
python-frontmatter>=1.0.0  # Markdown with metadata
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Analytics
pandas>=2.1.0
numpy>=1.26.0

# GitHub Integration
PyGithub>=2.1.0

# Date/Time
python-dateutil>=2.8.0

# Slack Integration
slack-sdk>=3.27.0

